docs(examples): rename issue resolver shared helper and remove grandfathered parity exception
