+++ examples/issue_resolver_common.py
