+++ tests/test_registry_docs.py
@@ -10,9 +10,6 @@ REPO_ROOT = Path(__file__).resolve().parent.parent
 SKILL_PATTERN = re.compile(r"`([\w-]+/[\w-]+)`")
 EXAMPLE_PATTERN = re.compile(r"^\|[^|]*`([\w-]+\.py)`", re.MULTILINE)
 
-# TODO: script examples/issue_resolver_github_context.py is a shared helper module and
-#       should be renamed to examples/issue_resolver_common.py (see #183).
-GRANDFATHERED_EXAMPLES: set[str] = {"issue_resolver_github_context.py"}
 
 
 def get_manifested_skills(skills_root: Path) -> set[str]:
@@ -45,7 +42,6 @@ def example_scripts() -> set[str]:
         path.name
         for path in examples_dir.glob("*.py")
         if not path.name.endswith("_common.py")
-        and path.name not in GRANDFATHERED_EXAMPLES
     }
